Move the method for adjusting font size to a protocol
