+++ Extensions/Today/TodayViewController.swift
@@ -26,7 +26,6 @@ class TodayViewController: UIViewController, NCWidgetProviding, TodayWidgetAppea
         let label = imageButton.label
         label.textColor = TodayUX.labelColor
         label.tintColor = TodayUX.labelColor
-        label.font = UIFont.preferredFont(forTextStyle: .body)
         label.adjustsFontForContentSizeCategory = true
         return imageButton
     }
@@ -71,7 +70,6 @@ class TodayViewController: UIViewController, NCWidgetProviding, TodayWidgetAppea
         let widgetView: UIView!
         self.extensionContext?.widgetLargestAvailableDisplayMode = .compact
         viewModel.setViewDelegate(todayViewDelegate: self)
-        NotificationCenter.default.addObserver(self, selector: #selector(preferredContentSizeChanged(_:)), name: UIContentSizeCategory.didChangeNotification, object: nil)
         let effectView: UIVisualEffectView
 
         if #available(iOS 13, *) {
@@ -102,41 +100,11 @@ class TodayViewController: UIViewController, NCWidgetProviding, TodayWidgetAppea
         } else {
             buttonStackView.removeArrangedSubview(openCopiedLinkButton)
         }
-        adjustFonts()
     }
 
     override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
         let edge = size.width * TodayUX.buttonsHorizontalMarginPercentage
         buttonStackView.layoutMargins = UIEdgeInsets(top: 0, left: edge, bottom: 0, right: edge)
-    }
-
-    @objc func preferredContentSizeChanged(_ notification: Notification) {
-        adjustFonts()
-    }
-
-    func adjustFonts() {
-        let size = traitCollection.preferredContentSizeCategory
-        switch size {
-        case let size where size >= .accessibilityMedium:
-            resize(size: 25)
-        case let size where size <= .extraExtraExtraLarge && size > .extraLarge:
-            resize(size: 15)
-        case let size where size >= .large && size <= .extraLarge:
-            resize(size: 14)
-        case let size where size == .medium:
-            resize(size: 12)
-        case let size where size >= .extraSmall && size <= .small:
-            resize(size: 8)
-        default:
-            resize(size: UIFont.systemFontSize)
-        }
-    }
-
-    func resize(size: CGFloat) {
-        newTabButton.label.font = newTabButton.label.font.withSize(size)
-        newPrivateTabButton.label.font = newPrivateTabButton.label.font.withSize(size)
-        openCopiedLinkButton.label.font = openCopiedLinkButton.label.font.withSize(size)
-        closePrivateTabsButton.label.font = closePrivateTabsButton.label.font.withSize(size)
     }
 
     func widgetMarginInsets(forProposedMarginInsets defaultMarginInsets: UIEdgeInsets) -> UIEdgeInsets {
